backend/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
qrcode==7.4.2
pandas==2.1.1
scikit-learn==1.3.1
PyJWT==2.8.0
Pillow==10.0.1

Flask-CORS==4.0.0
